+++ x/stake/types/pool.go
@@ -9,13 +9,8 @@ import (
 
 // Pool - dynamic parameters of the current state
 type Pool struct {
-	LooseTokens       int64   `json:"loose_tokens"`        // tokens not associated with any validator
-	UnbondedTokens    int64   `json:"unbonded_tokens"`     // reserve of unbonded tokens held with validators
-	UnbondingTokens   int64   `json:"unbonding_tokens"`    // tokens moving from bonded to unbonded pool
-	BondedTokens      int64   `json:"bonded_tokens"`       // reserve of bonded tokens
-	UnbondedShares    sdk.Rat `json:"unbonded_shares"`     // sum of all shares distributed for the Unbonded Pool
-	UnbondingShares   sdk.Rat `json:"unbonding_shares"`    // shares moving from Bonded to Unbonded Pool
-	BondedShares      sdk.Rat `json:"bonded_shares"`       // sum of all shares distributed for the Bonded Pool
+	LooseTokens       sdk.Rat `json:"loose_tokens"`        // tokens which are not bonded in a validator
+	BondedTokens      sdk.Rat `json:"bonded_tokens"`       // reserve of bonded tokens
 	InflationLastTime int64   `json:"inflation_last_time"` // block which the last inflation was processed // TODO make time
 	Inflation         sdk.Rat `json:"inflation"`           // current annual inflation rate
 
@@ -35,13 +30,8 @@ func (p Pool) Equal(p2 Pool) bool {
 // initial pool for testing
 func InitialPool() Pool {
 	return Pool{
-		LooseTokens:             0,
-		BondedTokens:            0,
-		UnbondingTokens:         0,
-		UnbondedTokens:          0,
-		BondedShares:            sdk.ZeroRat(),
-		UnbondingShares:         sdk.ZeroRat(),
-		UnbondedShares:          sdk.ZeroRat(),
+		LooseTokens:             sdk.ZeroRat(),
+		BondedTokens:            sdk.ZeroRat(),
 		InflationLastTime:       0,
 		Inflation:               sdk.NewRat(7, 100),
 		DateLastCommissionReset: 0,
@@ -52,108 +42,78 @@ func InitialPool() Pool {
 //____________________________________________________________________
 
 // Sum total of all staking tokens in the pool
-func (p Pool) TokenSupply() int64 {
-	return p.LooseTokens + p.UnbondedTokens + p.UnbondingTokens + p.BondedTokens
+func (p Pool) TokenSupply() sdk.Rat {
+	return p.LooseTokens.Add(p.BondedTokens)
 }
 
 //____________________________________________________________________
 
 // get the bond ratio of the global state
 func (p Pool) BondedRatio() sdk.Rat {
-	if p.TokenSupply() > 0 {
-		return sdk.NewRat(p.BondedTokens, p.TokenSupply())
+	supply := p.TokenSupply()
+	if supply.GT(sdk.ZeroRat()) {
+		return p.BondedTokens.Quo(supply)
 	}
 	return sdk.ZeroRat()
 }
 
-// get the exchange rate of bonded token per issued share
-func (p Pool) BondedShareExRate() sdk.Rat {
-	if p.BondedShares.IsZero() {
-		return sdk.OneRat()
+//_______________________________________________________________________
+
+func (p Pool) looseTokensToBonded(bondedTokens sdk.Rat) Pool {
+	p.BondedTokens = p.BondedTokens.Add(bondedTokens)
+	p.LooseTokens = p.LooseTokens.Sub(bondedTokens)
+	if p.LooseTokens.LT(sdk.ZeroRat()) {
+		panic(fmt.Sprintf("sanity check: loose tokens negative, pool: %v", p))
 	}
-	return sdk.NewRat(p.BondedTokens).Quo(p.BondedShares)
+	return p
 }
 
-// get the exchange rate of unbonding tokens held in validators per issued share
-func (p Pool) UnbondingShareExRate() sdk.Rat {
-	if p.UnbondingShares.IsZero() {
-		return sdk.OneRat()
+func (p Pool) bondedTokensToLoose(bondedTokens sdk.Rat) Pool {
+	p.BondedTokens = p.BondedTokens.Sub(bondedTokens)
+	p.LooseTokens = p.LooseTokens.Add(bondedTokens)
+	if p.BondedTokens.LT(sdk.ZeroRat()) {
+		panic(fmt.Sprintf("sanity check: bonded tokens negative, pool: %v", p))
 	}
-	return sdk.NewRat(p.UnbondingTokens).Quo(p.UnbondingShares)
-}
-
-// get the exchange rate of unbonded tokens held in validators per issued share
-func (p Pool) UnbondedShareExRate() sdk.Rat {
-	if p.UnbondedShares.IsZero() {
-		return sdk.OneRat()
-	}
-	return sdk.NewRat(p.UnbondedTokens).Quo(p.UnbondedShares)
+	return p
 }
 
 //_______________________________________________________________________
+// Inflation
 
-func (p Pool) addTokensUnbonded(amount int64) (p2 Pool, issuedShares PoolShares) {
-	issuedSharesAmount := sdk.NewRat(amount).Quo(p.UnbondedShareExRate()) // tokens * (shares/tokens)
-	p.UnbondedShares = p.UnbondedShares.Add(issuedSharesAmount)
-	p.UnbondedTokens += amount
-	p.LooseTokens -= amount
-	if p.LooseTokens < 0 {
-		panic(fmt.Sprintf("sanity check: loose tokens negative, pool: %v", p))
-	}
-	return p, NewUnbondedShares(issuedSharesAmount)
+const precision = 100000000000     // increased to this precision for accuracy
+var hrsPerYrRat = sdk.NewRat(8766) // as defined by a julian year of 365.25 days
+
+// process provisions for an hour period
+func (p Pool) ProcessProvisions(params Params) Pool {
+	p.Inflation = p.NextInflation(params)
+	provisions := p.Inflation.Mul(p.TokenSupply()).Quo(hrsPerYrRat)
+
+	// TODO add to the fees provisions
+	p.LooseTokens = p.LooseTokens.Add(provisions)
+	return p
 }
 
-func (p Pool) removeSharesUnbonded(shares sdk.Rat) (p2 Pool, removedTokens int64) {
-	removedTokens = p.UnbondedShareExRate().Mul(shares).RoundInt64() // (tokens/shares) * shares
-	p.UnbondedShares = p.UnbondedShares.Sub(shares)
-	p.UnbondedTokens -= removedTokens
-	p.LooseTokens += removedTokens
-	if p.UnbondedTokens < 0 {
-		panic(fmt.Sprintf("sanity check: unbonded tokens negative, pool: %v", p))
-	}
-	return p, removedTokens
-}
+// get the next inflation rate for the hour
+func (p Pool) NextInflation(params Params) (inflation sdk.Rat) {
 
-func (p Pool) addTokensUnbonding(amount int64) (p2 Pool, issuedShares PoolShares) {
-	issuedSharesAmount := sdk.NewRat(amount).Quo(p.UnbondingShareExRate()) // tokens * (shares/tokens)
-	p.UnbondingShares = p.UnbondingShares.Add(issuedSharesAmount)
-	p.UnbondingTokens += amount
-	p.LooseTokens -= amount
-	if p.LooseTokens < 0 {
-		panic(fmt.Sprintf("sanity check: loose tokens negative, pool: %v", p))
-	}
-	return p, NewUnbondingShares(issuedSharesAmount)
-}
+	// The target annual inflation rate is recalculated for each previsions cycle. The
+	// inflation is also subject to a rate change (positive or negative) depending on
+	// the distance from the desired ratio (67%). The maximum rate change possible is
+	// defined to be 13% per year, however the annual inflation is capped as between
+	// 7% and 20%.
 
-func (p Pool) removeSharesUnbonding(shares sdk.Rat) (p2 Pool, removedTokens int64) {
-	removedTokens = p.UnbondingShareExRate().Mul(shares).RoundInt64() // (tokens/shares) * shares
-	p.UnbondingShares = p.UnbondingShares.Sub(shares)
-	p.UnbondingTokens -= removedTokens
-	p.LooseTokens += removedTokens
-	if p.UnbondedTokens < 0 {
-		panic(fmt.Sprintf("sanity check: unbonding tokens negative, pool: %v", p))
-	}
-	return p, removedTokens
-}
+	// (1 - bondedRatio/GoalBonded) * InflationRateChange
+	inflationRateChangePerYear := sdk.OneRat().Sub(p.BondedRatio().Quo(params.GoalBonded)).Mul(params.InflationRateChange)
+	inflationRateChange := inflationRateChangePerYear.Quo(hrsPerYrRat)
 
-func (p Pool) addTokensBonded(amount int64) (p2 Pool, issuedShares PoolShares) {
-	issuedSharesAmount := sdk.NewRat(amount).Quo(p.BondedShareExRate()) // tokens * (shares/tokens)
-	p.BondedShares = p.BondedShares.Add(issuedSharesAmount)
-	p.BondedTokens += amount
-	p.LooseTokens -= amount
-	if p.LooseTokens < 0 {
-		panic(fmt.Sprintf("sanity check: loose tokens negative, pool: %v", p))
+	// increase the new annual inflation for this next cycle
+	inflation = p.Inflation.Add(inflationRateChange)
+	if inflation.GT(params.InflationMax) {
+		inflation = params.InflationMax
 	}
-	return p, NewBondedShares(issuedSharesAmount)
-}
+	if inflation.LT(params.InflationMin) {
+		inflation = params.InflationMin
+	}
 
-func (p Pool) removeSharesBonded(shares sdk.Rat) (p2 Pool, removedTokens int64) {
-	removedTokens = p.BondedShareExRate().Mul(shares).RoundInt64() // (tokens/shares) * shares
-	p.BondedShares = p.BondedShares.Sub(shares)
-	p.BondedTokens -= removedTokens
-	p.LooseTokens += removedTokens
-	if p.UnbondedTokens < 0 {
-		panic(fmt.Sprintf("sanity check: bonded tokens negative, pool: %v", p))
-	}
-	return p, removedTokens
+	return inflation.Round(precision)
 }